workaround check of inet_{ntop,pton} for bsdi4.
bsdi4 ships with BIND8 resolver, and the way it defines inet_{ntop,pton}
is rather strange; it does not ship with libc symbol "inet_ntop".
it ships with "_inet_ntop", and has a #define macro in header to rename it.
AC_CHECK_FUNCS will always fail for resolver functions.

i consider it as a bug in merger.  bsdi should have done the following:
- supply _inet_ntop for libc-internal calls
- supply weak symbol inet_ntop for external reference
- no #define in headers

from kame.