Don't print the value of a RADIUS attribute if the length is less than or equal to 2 (the length includes the two bytes of type and length).
Don't print the values of RADIUS attributes that are supposed to be of a
particular length if they have a different length.

Don't use the result of "ctime()", unmodified, when printing an
attribute value; the string "ctime()" supplies includes a newline, which
we need to remove.

The "PRINT_HEX()" macro expands into code that modifies its first
argument; always set the variable in question to the appropriate value
before calling the macro.