If we run out of data in the when processing tagged parameters, return "success", not "failure" - there's no special "end of list" parameter. Get rid of the check for a parameter tag of 0xff, for the same reason.
Take out an extra space between the "[" and the list of rates in a
"supported rates" tagged parameter, and don't print the "[" or the "
Mbit]", or the leading space if there are no supported rates (or there
is no "supported rates" parameter.

Use "fn_print" to print the SSID, in case there are non-printable
characters in it.

Show the channel number in decimal, not hex, as that's how it's given in
the IEEE 802.11 spec.

Show the supported rates in a probe response, if present.

Get rid of random extra trailing space in tagged parameter output.