Fix the ATMARP dissector to correctly dissect RFC 2225 ATM ARP.
Fix the ARP dissector to check the hardware type before checking whether
the addresses are in the captured data, and call the ATMARP dissector if
the hardware type is 19.  Also fix it to return after the ATMARP
dissector returns.

Use "TTEST2()" to check whether the addresses are in the captured data.