Constify.
Check for zero-length items in "isakmp_sub0_print()", and return NULL if
one is seen.  Have "isakmp_sub_print()" check for that and break out of
the loop and return NULL if "isakmp_sub0_print()" returns NULL, so we
don't loop forever printing the same item.  Have callers of
"isakmp_sub_print()" check for NULL and return NULL as well.

Have "isakmp_sub_print()" take the depth as an argument, rather than
have it as a static variable, for cleanliness.