Get rid of an unused variable.
Cast lengths to "int" when using them in "%*" formats.

Restore the "TODO print UTF-8 decoded text" comment, as RFC 2516 says
that text in PPPoE is, in fact, UTF-8, and we could, in principle, do
more than just dumping stuff as hex if there are more non-printable than
printable characters (and dumping it as text if there aren't).