Updates from Bruce Simpson:
	Fix numerous bugs in my AODV submission, red faces all round,
	mainly due to misunderstanding of EXTRACT_STRING() macro
	arguments (because these are macros, type checking didn't
	happen).

	AODV HELLO extensions are now correctly printed.

	The code is closer to FreeBSD's style guidelines.

Don't use "u_char" and "u_long", use "u_int8_t" and "u_int32_t".
"u_long" is definitely wrong as it's 64 bits on some platforms; "u_char"
is safe, but the official tcpdump style uses "u_int8_t".

At least according to the (long-since-expired) draft-perkins-aodv6-01,
the V6 AODV messages have different message type values from the V4 ones
in draft-ietf-manet-aodv-13.  However, draft-ietf-manet-aodv-13 says
that the only differences are in the addresses; we support both.

Clean up the way we tell the AODV6 dissector that we're running atop
IPv6 a bit.

Use "udpipaddr_print()" if we're printing AODV packets as a result of
saying "anything that matches this filter is AODV", rather than doing it
ourselves under the assumption that it's running over IPv4.