"icmp6_rrenum_print()" doesn't use its "icmp6len" argument (and, as that just reflects the payload length, the snapshot length has been trimmed to correspond to it, so it doesn't need to check it merely to avoid running past the end of the IPv6 payload); get rid of it.
Pass "icmp6_print()" the IPv6 payload length (i.e., the length of the
ICMPv6 packet), rather than having it compre the length itself.

Make the length argument to "icmp6_cksum()" unsigned, to match the value
passed to it.

Fix the lengths passed to "sctp_print()", "tcp_print()", and
"udp_print()" (we update "len" in the header-processing loop to
correspond to the remaining payload, so we just pass it to those
routines).