Note that Mac OS X, as one of the BSDs, has SIGINFO; note that Digital/Tru64 UNIX has it as well (although the tty(7) man page doesn't mention VSTATUS - stty(1) mentions "status", however, and signal(4) mentions SIGINFO, and some Tru64->HP-UX transition pages mention SIGINFO as a feature available in Tru64 but not HP-UX).
Note that on some systems (e.g., Mac OS X), you might have to set your
"status" character as it defaults, on those systems, to "not set".

Get rid of the reference to "traffic(1C)" (it's not referred to in the
man page, and about the only connection it has with tcpdump is that they
both watch network traffic; traffic(1C) is a SunOS 4.x-ism, not present
on other platforms), and add a reference to pfconfig(8) for
Digital/Tru64 (it *is* referred to, in the section of what privileges
you need in order to capture traffic).