Some BSDs define IPPROTO_IGRP as 88 and don't define IPPROTO_EIGRP, so we also define it as 88, so we can't have
	case IPPROTO_IGRP:
	case IPPROTO_EIGRP:

in a switch statement.

The current IANA protocol number assignments page lists 9 as "any
private interior gateway (used by Cisco for their IGRP)" and 88 as
"EIGRP" from Cisco, so we define IPPROTO_PIGP as 9 (which matches what
those BSDs do) and IPPROTO_EIGRP as 88 (so we don't bother with
IPPROTO_IGRP at all), and do

	case IPPROTO_PIGP:
	case IPPROTO_EIGRP:

It's not clear that we should handle both with the same print routine,
however - Cisco IGRP and Cisco EIGRP don't appear to have the same
packet format.