Rename "wcnt" in some routines to "wct", just as it's called in "print_smb()".
Don't print "smb_vwv[] =" - just print the "smb_vwv[i] =" for each of
the words in the word data if we're dumping it as raw data.

Only print "smb_buf[] =" if we're dumping the byte data as raw data.

Always print the byte count (except for Trans2, but we should probably
print it there as well).