Check for a negative "advance" value, which indicates an error.
Don't make the last statement in a loop a break statement, as that means
the body of the loop will be executed at most once.

When looping over the data in a TCP segment containing BGP packets, loop
until we reach the first of the end of the captured data or the end of
the segment, not just the end of the captured data.