Make "subtype_text[]" have 16 elements, for all 16 possible subtypes, and make all unknown subtypes have null pointers and, if the pointer for a subtype is null, print the subtype as "Unknown subtype".
Compute the sizes of the tables of authentication algorithm names,
status code strings, and reason code strings at compile time, and use
those values to check whether the value is in the table or not.  Get rid
of trailing NULL pointers in those tables.

Fix a typo in one reason code string.