Don't set the length of the attributes based on the snapshot length, just add some additional TCHECK/TCHECK2 bounds checks to the code that dissects attributes and let that handle the snapshot length checks.
Do the length check once per attribute, rather than doing a single check
up front.

Use TCHECK/TCHECK2 and TTEST/TTEST2, so that we print "too short"
indications.  Make the "too short" indications all look the same.

Rename "radius_attr_print()" to "radius_attrs_print()" to make it
clearer that it has a loop to print all attributes, rather than just
printing one attribute.

Add in a bunch of additional bounary/length checks, and do the

As per Steiner Haug, the length of a vendor-specific attribute includes
the type and length bytes, so subtract two from the length to get the
length of the attribute's data.