FILES isn't used to make the tarball any more; get rid of it.
Update HDR in Makefile.in to include all headers.

Add an EXTRA_DIST variable to include all the files that should go into
the tarball and that aren't in CSRC or HDR, other than tcpdump.1.

Use CSRC, HDR, and EXTRA_DIST to determine what goes into the tarball,
along with tcpdump.1, rather than doing a "make distclean" and putting
everything into the tarball; that way, you can do "make releasetar"
without cleaning out the current directory.

Given that we're not just tarring up the entire source directory, we can
make the tarball directory as a subdirectory of the current directory and
put the tarball into the current directory, rather than putting it into
a (not-entirely-obvious) ../n directory.  Clean out the tarball
directory when we're done.