Don't clear stuff we'll be overwriting in its entirety.
We only need to clear out the tha structure if we support IPv6, so that
the addresses are IPv6 addresses, but we only have IPv4 addresses, so
that the rest of the address bytes need to be set somehow.

Also, explain what we're doing, so people don't ask why we're not
zeroing it if we don't support IPv6.