Handle UDP checksums more like TCP checksums.
Instead of printing the result of udp_cksum() if it's non-zero, print
the checksum field value and the value it should have had.  That means
that what we print is the same regardless of whether we're running on a
big-endian or little-endian machine.

Also, just as we did with TCP:

Check -v and -K, and the fragmented flag, up front; then check the IP
version etc..  Don't check for IPv6 if we already know it's IPv4.  Fetch
the checksum field only once.

Update some test files for the new output format.