Use PATH_MAX, not NAME_MAX.
NAME_MAX is the maximum length of a file pathname *component*; PATH_MAX
is the maximum length of a file pathname.  We're dealing with pathnames,
so use PATH_MAX.  (On some systems, NAME_MAX can be as low as 14,
presumably for binary compatibility with V7-era data structures
containing file names, as there are probably few if any non-historic
UN*Xes around with 14-character file name limitations.)

Boost the default PATH_MAX to 1024 while we're at it.